Vivado 2015.4: IPs migrated, errors in build fixed
the project and scripts have been changed to allow the logic to be opened and
changed with Vivado 2015.4 (the current supported version on master).
during the Makefile build, the right project environment will be detected
automatically depending on the settings64.sh that has been executed before.

notice: due to changes in the Xilinx SDK, the Makefile build will fail for the
2015.4 environment right after the bitstream and hardware definitions have been
generated.